Initial conversion to GameClass/GameState system.
The RpgGame GameClass represents theoverall game and any global constants
or variables that need to be shared throughout. It handles the basic update
loop, as well as controlling the the individual GameStates.

A GameState is a unique segment of the game that has vastly different
input, logic updates, and rendering needs. Switching between GameStates
can easily be done through the GameClass.

TestState was added demonstrate how our existing functionality would be
represented in GameState form.