Levels now save and load from a file.
All data associated with a level can now be saved/loaded in the a '.lvl' format.
We no longer need to manually load the Level TileSet as that info is now
stored within the level file itself and loaded when the level is loaded.

Levels now automatically save when leaving the Editor state.