Add auto Terrain generation to the Editor.
The Editor now has a special "Terrain" mode that will autogenerate tiles
based on their proximity to tiles of the same "terrain" type. Placing tiles
in this mode will also update neighbor tiles so they "blend" correctly. Be
careful, as this could mess up manually placed tiles. To active Terrain mode,
just select one of the new Terrain Icons to the right of the screen below
the normal TileSet UI. Selecting a normal TileSet tile will deactive the
mode.

The basic TileSet class had to be extended (AdvancedTileSet) to allow for
meta information to be stored/loaded in a new tileset file format. This
format contains all the terrain data for auto generation to work.

Refs:
http://gamedevelopment.tutsplus.com/tutorials/how-to-use-tile-bitmasking-to-auto-tile-your-level-layouts--cms-25673